requirements.txt
PyCryptodome
semver
